refactor(modals.WireframeVersionNotice): migrate translations location to adjacent
