--- src/lib/shared/services/i18n/translations/modals/wireframe-version-notice/vi.ts
@@ -1,11 +1,0 @@
-export const vi = {
-  title: 'Đang thi công',
-  description:
-    'Bạn đang xem bản mẫu cho trang web chính thức của Svelte Vietnam. Dự án đang trong giai đoạn thiết kế và phát triển.<br/>Để biết thêm thông tin, hãy xem qua <a href={GITHUB_LINKS.PROJECT_REFERENCES} class="c-link" target="_blank" rel="noreferrer">Project References</a>. Theo dõi tiến độ và lộ trình tại <a href={GITHUB_LINKS.PROJECT} class="c-link" target="_blank" rel="noreferrer">project sveltevietnam.dev</a>.',
-  feedbacks:
-    'Mọi <a href={GITHUB_LINKS.DISCUSSION} class="c-link" target="_blank" rel="noreferrer">phản hồi và đóng góp</a > đều được hoan nghênh. Xin cảm ơn!',
-  ctas: {
-    continue: 'Tiếp tục',
-    doNotShowAgain: 'Không hiển thị lại',
-  },
-};
